requirements.txt
easyocr
opencv-python
streamlit
numpy
pillow
pyttsx3
deep-translator
